Moved timing function to a separate file
